refactor: observe 로직 분리
